feat: add dynamic trimesh example
